self-checkout-system/requirements.txt
numpy
opencv-python
PyYAML
Flask
kafka-python
pytest